\name{Aca3Turns-package}
\alias{Aca3Turns-package}
\alias{Aca3Turns}
\docType{package}
\title{
\packageTitle{Aca3Turns}
}
\description{
\packageDescription{Aca3Turns}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Aca3Turns}
\packageIndices{Aca3Turns}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Aca3Turns}

Maintainer: \packageMaintainer{Aca3Turns}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
